アイテム追加 ロジックの分離
